commit_templater: function to present change ids in readable mixed case
This adds a `readable_case` function and some tests to give examples.

This focuses on the `k-z` part of the alphabet that's used for change ids. The
normal hexadecimal alphabet 0-9a-f doesn't have many confusable letters, except
perhaps `b` and `d`.

## Backstory and design

This is the third attempt at making change ids readable at a glance. Previously,
I was suggesting printing them in uppercase (which I use in my personal config,
but several people disliked the idea since they perceive UPPERCASE as yelling)
or printing half the alphabet in uppercase.

Since then, I found that uppercase doesn't solve all of my issues either, as
it's quite easy to confuse `OQ` or `UV` when reading quickly.

So, I now have a variation of the second idea where I carefully pick which
letters to capitalize. See comments and test in the commit for examples.