test_conflicts: Demo bug with conflicts in files with conflict markers
If a conflict happens between file contents that itself contains conflict
markers, jj can (unsurprisingly) get confused.

This is the kind of bug that sounds serious in theory, but might not matter in
practice.

This bug seems like a bit of a headache to fix completely. For example, one fix
would be to start conflicts with a longer sequence of `<<<<<<<<<` if the
conflict contains conflict markers. However, it's then unclear to me what to do
if a file contains `<<<<<<<` outside a conflict. Perhaps `jj` could create a
trivial conflict in this case, looking like:

<<<<<<<<<<<<
++++++++++++
<<<<<<<
%%%%%%%%%%%%
>>>>>>>>>>>>