lib `file`: Refactor merge of file hunks
The goal here is to create a single operation to split a conflicted file text into a sequence of (potentially conflicted) hunks, both for regular conflicts and conflict diffs.

The new algorithm is not efficient and I'm guessing an algorithm that gives better results (more logical hunks) is possilbe.

We may want to instead keep the conflicts and conflict diff code path separate until the algorithm improves, let me know.