templates: make `op_summary` more readable and colorful
This is meant to be a quick fix before I update the demos in the README.
The primary reason to update them is the branch -> bookmark transition,
but I noticed that `jj undo`'s output is currently not very pretty.

I believe printing both the start and end times is excessive for a
summary. For a long-running operation like `jj describe` with the editor
open for a long time, the ending time seems more meaningful.

I was a bit torn on whether to use `format_timestamp(self.time().end())`
or `self.time().end().ago()`. The latter looks better, but is less
configurable and worse for dates long ago. In the future, we could add a
`format_op_summary_timestamp` function and/or a template function that
uses `.ago()` for recent dates and absolute dates for old dates.