cli `evolog`: new `--diff-snapshots` option for use with `-p`
Currently, `jj evolog -p` shows *interdiffs* (as in `jj interdiff`)
between different versions of the change, so that a rebase does not
result in numerous changes shown in the output. This is often useful,
but not always.

This adds the option of `jj evolog -p --diff-snapshots` that changes
this behavior to show normal diffs between versions of the change.

One example where this could be useful: let's say you do `jj abandon @-`
and you later want to do `jj diffedit --to @ --from
version_of_@-_before_abandon`. Currently, running `jj evolog -p` will
not help you find the commit id for `version_of_@-_before_abandon`. With
`--diff-snapshots`, it will.

Another motivation for this is that, as I was working of `jj squash
--restore-target-descendants`, `jj evolog` of children of the target
commit showed changes. I think it'll be helpful to have an option that
gives the intuitive result of not showing any changes in this case.