cli: rename `--restore-descendants` to `--preserve-descendant-contents` AKA `--pdc`
Fixes https://github.com/jj-vcs/jj/discussions/5951

There have also been discussions on Discord on and off; people seem to
agree that `--preserve-descendant-contents` is clear if verbose. I hope
that a short alias might address that (though I don't intent to make a
habit of having many cryptic short aliases), though we could also decide
to skip it for now and see whether it's really needed.

Some other options considered:

- `keep-descendant-contents`. If all else were equal, I like it better,
but it'd be
  confusingly similar to `jj squash --keep-emptied` once `jj squash
  --keep-descendant-contents` was implemented, and would also conflict
  with its tab completion.
- `verbatim-descendants`
- `reparent-descendants`

The last two look a bit cryptic, depending on the person.

Note that `jj rebase` will also need a `--preserve-contents` flag for
preserving the contents of the commit actually being rebased. I'm not
sure whether that should also be `--pc` or not.