tests: make `git push --named` tests consistent again
Follows up on https://github.com/jj-vcs/jj/pull/5698 and
https://github.com/jj-vcs/jj/pull/6236.

The behavior of `--named` in corner cases differs. This wasn't detected
in #5678 because of a bug in the test. This commit better illustrates
the more important case where the behavior is the same, and the less
important case where it differs.