split: the commit with the remaining changes keeps the change id
The commit with the selected changes gets a new change id.
The commit with the remaining changes (and usually most of the changes)
keeps the change id and the attached bookmarks.

This no-implicit-move behavior aligns with other jj commands.

This is consistent with the behavior when using the -A/-B/-d flags:
before this change, when not using those flags, the remaining changes
were getting the new change id, making the behavior different when
using those flags or when not using it.