lib squash_commits: Fix partial squash into parent, forbid unsupported cases
The behavior in the parent commit would be quite unexpected for a user squashing
a file into the parent, and is generally confusing. This changes the behavior,
and forbids cases where the new behavior doesn't make sense.

Another option would be to forbid partial squash with
`--restore-descendants` entirely. I do, however, suspect it could be
useful relatively often in the case where you squash into a parent.

See inline comment for more motivation. Before this commit, the behavior
was the II from the comment AFAICT (but which behavior is easiest to
implement could change if `rewrite_commits` was refactored to do things
in a different order).