cli `diff`: warn if `-r` arg does not contain `::` or `..` and expands to multiple revisions
I think it's better to do this only when the user passed in a *single*
-r arg. If the user does `jj diff -r a -r b`, they probably expect
to get multiple revisions.

This supercedes #6931 because of
https://github.com/jj-vcs/jj/pull/6931#discussion_r2217975635

I previosly considered actually checking that the revision cannot be
expressed as (say) `a..b` for any `a`, `b`, but I think that's both
harder and more confusing for users. Textual check is less clever,
but should be pretty intuitive.

Fixes #6723