testutils: add `git` feature, make gix dependency optional
This makes it possible to use `testutils` without depending on all of
`gix`.

Users of testutils without the Git feature would currently either have
to use the LocalBackend or provide their own replacement to `TestRepo`.

`jj-lib`'s test still depend on Git. Both before and after this PR,
`cargo test --no-build -p jj-lib --no-default-features` fails to
compile. This seems like a difficult problem to address; this PR merely
tries to not make the situation worse.

**Motivation**: In at least one ERSC use-case, we'd like to use `testutils`, but
the gix dependency makes tests crash with the following error:

```
[test]     thread 'gitoxide.in_parallel.produce.0' panicked at code/...:
[test]     there is no reactor running, must be called from the context of a Tokio 1.x runtime
```

It's not clear to me why this happens (@davidbarsky claims to have an
idea), but this PR seems like a good workaround that also decreases the
amount of things we'd need to compile.